requirements.txt
ollama==0.5.1
python-telegram-bot==22.3
loguru==0.7.2
